llm-server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
numpy==1.26.4
python-dotenv==1.0.1
google-generativeai==0.7.0